refactor(tests): Move `Richtext` Fixture to example-data
`Richtext` fixture in itest was quite matured, providing
attribute-support as well as documentation of possible
elements, which may be contained in one element.

Moved these methods to new example-data and replaced
or enhanced existing RichText methods in example-data.
